added Games from This Month
Under the heading "From This Month", on the home page, in the sidebar,
are listed games submitted in the current month, both for this year and
all past years. It's pretty much intended to keep our rich history from
being forgotten.

Come to think of it, it might be worthwhile to turn this on for past
Featured Events.