Made Facebook link post metadata autodetection friendlier
The entire-website image and description are now only applied to the
front page, so the description "A student organization of game makers
and fans […]" is only applied if you link the home page, not if you
link, say, a game. Same goes for the logo image being used for the
thumbnail.

The logo image should still be scraped if the page in question doesn't
have an existing image, though.

Hacky hacky hack!

(The correct way to do all this is to automatically populate ‘<meta
property="og:’ tags with information from each content type (especially
games), but that requires installing another module or writing more
code, both of which are big investments in time and instability.
Pragmatism!™)